Enable the M999 code to unlock the machine.
* Modify the class ScreenEmergency to not lock the machine completely:
  - The draw() method changes to allow the machine to read the serial incoming commands.
  - The constructor sets the stop_buffer related flags to only handle the M999 command.
* The method lcd_emergency_stop() in GuiManager is not static anymore.
* Fix some includes on Marlin_main file.

+++ Marlin/GuiAction.cpp
@@ -17,7 +17,7 @@ extern bool cancel_heatup;
 extern bool stop_planner_buffer;
 extern bool stop_buffer;
 bool change_filament = false; 
-extern int stop_buffer_code;
+extern uint16_t stop_buffer_code;
 
 static float manual_feedrate[] = MANUAL_FEEDRATE;
